refactor: hyozaparent 코드 리팩토링 #7
